requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.6
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
psycopg2-binary==2.9.5
postgres==4.0
pytest-django==4.5.2